<fix>[zstack]: cast math operands before assignment
math operands should be cast before assignment

Related: ZSTAC-53781

+++ compute/src/main/java/org/zstack/compute/vm/VmInstanceBase.java
@@ -5219,7 +5219,7 @@ public class VmInstanceBase extends AbstractVmInstance {
             @Override
             public void run(FlowTrigger chain, Map data) {
                 DesignatedAllocateHostMsg msg = new DesignatedAllocateHostMsg();
-                msg.setCpuCapacity(cpuNum - oldCpuNum);
+                msg.setCpuCapacity((long) cpuNum - oldCpuNum);
                 msg.setMemoryCapacity(struct.alignedMemory - oldMemorySize);
                 msg.setOldMemoryCapacity(oldMemorySize);
                 msg.setAllocatorStrategy(HostAllocatorConstant.DESIGNATED_HOST_ALLOCATOR_STRATEGY_TYPE);
@@ -5252,7 +5252,7 @@ public class VmInstanceBase extends AbstractVmInstance {
             public void rollback(FlowRollback chain, Map data) {
                 if (result) {
                     ReturnHostCapacityMsg msg = new ReturnHostCapacityMsg();
-                    msg.setCpuCapacity(cpuNum - oldCpuNum);
+                    msg.setCpuCapacity((long) cpuNum - oldCpuNum);
                     msg.setMemoryCapacity(struct.alignedMemory - oldMemorySize);
                     msg.setHostUuid(self.getHostUuid());
                     msg.setServiceId(bus.makeLocalServiceId(HostAllocatorConstant.SERVICE_ID));
